+++ scripts/utilities/string.rb
@@ -27,5 +27,6 @@ class String
   def light_blue
     colorize(36)
   end
+
   alias :cyan :light_blue
 end